+++ opentelemetry-proto/src/transform/metrics.rs
@@ -208,17 +208,16 @@ pub mod tonic {
         fn from(hist: &SdkHistogram<T>) -> Self {
             TonicHistogram {
                 data_points: hist
-                    .data_points
-                    .iter()
+                    .data_points()
                     .map(|dp| TonicHistogramDataPoint {
-                        attributes: dp.attributes.iter().map(Into::into).collect(),
+                        attributes: dp.attributes().map(Into::into).collect(),
                         start_time_unix_nano: to_nanos(hist.start_time),
                         time_unix_nano: to_nanos(hist.time),
                         count: dp.count,
                         sum: Some(dp.sum.into_f64()),
                         bucket_counts: dp.bucket_counts.clone(),
                         explicit_bounds: dp.bounds.clone(),
-                        exemplars: dp.exemplars.iter().map(Into::into).collect(),
+                        exemplars: dp.exemplars().map(Into::into).collect(),
                         flags: TonicDataPointFlags::default() as u32,
                         min: dp.min.map(Numeric::into_f64),
                         max: dp.max.map(Numeric::into_f64),
@@ -236,10 +235,9 @@ pub mod tonic {
         fn from(hist: &SdkExponentialHistogram<T>) -> Self {
             TonicExponentialHistogram {
                 data_points: hist
-                    .data_points
-                    .iter()
+                    .data_points()
                     .map(|dp| TonicExponentialHistogramDataPoint {
-                        attributes: dp.attributes.iter().map(Into::into).collect(),
+                        attributes: dp.attributes().map(Into::into).collect(),
                         start_time_unix_nano: to_nanos(hist.start_time),
                         time_unix_nano: to_nanos(hist.time),
                         count: dp.count as u64,
@@ -255,7 +253,7 @@ pub mod tonic {
                             bucket_counts: dp.negative_bucket.counts.clone(),
                         }),
                         flags: TonicDataPointFlags::default() as u32,
-                        exemplars: dp.exemplars.iter().map(Into::into).collect(),
+                        exemplars: dp.exemplars().map(Into::into).collect(),
                         min: dp.min.map(Numeric::into_f64),
                         max: dp.max.map(Numeric::into_f64),
                         zero_threshold: dp.zero_threshold,
@@ -275,10 +273,10 @@ pub mod tonic {
                 data_points: sum
                     .data_points()
                     .map(|dp| TonicNumberDataPoint {
-                        attributes: dp.attributes.iter().map(Into::into).collect(),
+                        attributes: dp.attributes().map(Into::into).collect(),
                         start_time_unix_nano: to_nanos(sum.start_time),
                         time_unix_nano: to_nanos(sum.time),
-                        exemplars: dp.exemplars.iter().map(Into::into).collect(),
+                        exemplars: dp.exemplars().map(Into::into).collect(),
                         flags: TonicDataPointFlags::default() as u32,
                         value: Some(dp.value.into()),
                     })
@@ -296,13 +294,12 @@ pub mod tonic {
         fn from(gauge: &SdkGauge<T>) -> Self {
             TonicGauge {
                 data_points: gauge
-                    .data_points
-                    .iter()
+                    .data_points()
                     .map(|dp| TonicNumberDataPoint {
-                        attributes: dp.attributes.iter().map(Into::into).collect(),
+                        attributes: dp.attributes().map(Into::into).collect(),
                         start_time_unix_nano: gauge.start_time.map(to_nanos).unwrap_or_default(),
                         time_unix_nano: to_nanos(gauge.time),
-                        exemplars: dp.exemplars.iter().map(Into::into).collect(),
+                        exemplars: dp.exemplars().map(Into::into).collect(),
                         flags: TonicDataPointFlags::default() as u32,
                         value: Some(dp.value.into()),
                     })
@@ -318,8 +315,7 @@ pub mod tonic {
         fn from(ex: &SdkExemplar<T>) -> Self {
             TonicExemplar {
                 filtered_attributes: ex
-                    .filtered_attributes
-                    .iter()
+                    .filtered_attributes()
                     .map(|kv| (&kv.key, &kv.value).into())
                     .collect(),
                 time_unix_nano: to_nanos(ex.time),